Move tests to integration.rs
