+++ ref-tests/tests/ic-ref.rs
@@ -770,124 +770,3 @@ mod extras {
         })
     }
 }
-
-mod sign_send {
-    use crate::universal_canister::payload;
-    use ic_agent::agent::{
-        signed_query_inspect, signed_request_status_inspect, signed_update_inspect,
-    };
-    use ic_agent::agent::{Replied, RequestStatusResponse};
-    use ic_agent::AgentError;
-    use ref_tests::with_universal_canister;
-    use std::{thread, time};
-
-    #[ignore]
-    #[test]
-    fn query() {
-        with_universal_canister(|agent, canister_id| async move {
-            let arg = payload().reply_data(b"hello").build();
-            let signed_query = agent.query(&canister_id, "query").with_arg(arg).sign()?;
-
-            assert!(signed_query_inspect(
-                signed_query.sender,
-                signed_query.canister_id,
-                &signed_query.method_name,
-                &signed_query.arg,
-                signed_query.ingress_expiry,
-                signed_query.signed_query.clone()
-            )
-            .is_ok());
-
-            let result = agent
-                .query_signed(
-                    signed_query.effective_canister_id,
-                    signed_query.signed_query,
-                )
-                .await?;
-
-            assert_eq!(result, b"hello");
-            Ok(())
-        })
-    }
-
-    #[ignore]
-    #[test]
-    fn update_then_request_status() {
-        with_universal_canister(|agent, canister_id| async move {
-            let arg = payload().reply_data(b"hello").build();
-            let signed_update = agent.update(&canister_id, "update").with_arg(arg).sign()?;
-
-            assert!(signed_update_inspect(
-                signed_update.sender,
-                signed_update.canister_id,
-                &signed_update.method_name,
-                &signed_update.arg,
-                signed_update.ingress_expiry,
-                signed_update.signed_update.clone()
-            )
-            .is_ok());
-
-            let signed_request_status = agent.sign_request_status(
-                signed_update.effective_canister_id,
-                signed_update.request_id,
-            )?;
-
-            assert!(signed_request_status_inspect(
-                signed_request_status.sender,
-                &signed_request_status.request_id,
-                signed_request_status.ingress_expiry,
-                signed_request_status.signed_request_status.clone()
-            )
-            .is_ok());
-
-            let _request_id = agent
-                .update_signed(
-                    signed_update.effective_canister_id,
-                    signed_update.signed_update,
-                )
-                .await?;
-
-            let ten_secs = time::Duration::from_secs(10);
-            thread::sleep(ten_secs);
-
-            let response = agent
-                .request_status_signed(
-                    &signed_request_status.request_id,
-                    signed_request_status.effective_canister_id,
-                    signed_request_status.signed_request_status.clone(),
-                )
-                .await?;
-
-            assert!(
-                matches!(response, RequestStatusResponse::Replied{reply: Replied::CallReplied(result)} if result == b"hello")
-            );
-            Ok(())
-        })
-    }
-
-    #[ignore]
-    #[test]
-    fn forged_query() {
-        with_universal_canister(|agent, canister_id| async move {
-            let arg = payload().reply_data(b"hello").build();
-            let mut signed_query = agent.query(&canister_id, "query").with_arg(arg).sign()?;
-
-            signed_query.method_name = "non_query".to_string();
-
-            let result = signed_query_inspect(
-                signed_query.sender,
-                signed_query.canister_id,
-                &signed_query.method_name,
-                &signed_query.arg,
-                signed_query.ingress_expiry,
-                signed_query.signed_query.clone(),
-            );
-
-            assert!(matches!(result,
-                    Err(AgentError::CallDataMismatch{field, value_arg, value_cbor})
-                    if field == *"method_name" && value_arg == *"non_query" && value_cbor == *"query"));
-
-            Ok(())
-        })
-    }
-}
